Image button at home screen to generate list, ERRORS
Added a button imager at home to generate list. List view can be found in XML and the generatelist (oncreate) function can be found in MainActivity.

Current issue is that there is multiple oncreate initalizations but that shouldnt be a problem. Tried not using Listactivity but I am pretty sure its required. Will continue to debug